plane positionning relative to the world

+++ Scripts/Coord/Cartesian.cs
@@ -10,14 +10,14 @@ namespace Scripts.Coord
         /// <summary>
         /// Coordinates of a Cartesian instance
         /// </summary>
-        public int x, y;
+        public float x, y;
 
         /// <summary>
         /// Cartesian is a custructor of a cartesian system coordinate
         /// </summary>
         /// <param name="xPos"></param>
         /// <param name="yPos"></param>
-        public Cartesian(int x, int y)
+        public Cartesian(float x, float y)
         {
             this.x = x;
             this.y = y;
@@ -28,7 +28,7 @@ namespace Scripts.Coord
         /// </summary>
         /// <param name="c"></param>
         /// <returns>x or y coordinate, -1 if c isn't valid</returns>
-        public int Coord(int c)
+        public float Coord(int c)
         {
             if (c == 0) { return x; }
             else if (c == 1) { return y; }
